update button block structure, add render.php template
