+++ src/blocks/button/_bu-blocks-block-button-base.scss
@@ -8,12 +8,12 @@
 //
 //
 
-a:active,
-a:visited { // TO DO: move to global location
-	color: inherit;
-}
+// a:active,
+// a:visited { // TO DO: move to global location
+// 	color: inherit;
+// }
 
-.wp-block-button {
+.wp-block-bu-button {
 	@include transition( all 0.2s ease-in-out 0s);
 	display: inline-block;
 	background: transparent;
