+++ src/blocks/button/_bu-blocks-block-button-themeable.scss
@@ -61,7 +61,7 @@ $scope: blocks-get( 'scope' );
 		color: inherit;
 	}
 
-	&.wp-block-button {
+	&.wp-block-bu-button {
 		border-radius: blocks-get( 'border-radius' );
 		font-family: blocks-get( 'font-family-tertiary' );
 
@@ -70,7 +70,7 @@ $scope: blocks-get( 'scope' );
 		}
 	}
 
-	&.wp-block-button.is-style-default {
+	&.wp-block-bu-button.is-style-default {
 		&.has-dark-theme {
 			background: blocks-get( 'background-color-dark' );
 		}
@@ -80,7 +80,7 @@ $scope: blocks-get( 'scope' );
 		}
 	}
 
-	&.wp-block-button.is-style-outline {
+	&.wp-block-bu-button.is-style-outline {
 		border-color: blocks-get( 'color-secondary' );
 
 		&:hover {
@@ -96,7 +96,7 @@ $scope: blocks-get( 'scope' );
 		}
 	}
 
-	&.wp-block-button.has-primary-theme {
+	&.wp-block-bu-button.has-primary-theme {
 		&.is-style-default {
 			background: blocks-get( 'color-primary' );
 
@@ -123,7 +123,7 @@ $scope: blocks-get( 'scope' );
 		}
 	}
 
-	&.wp-block-button.has-secondary-theme {
+	&.wp-block-bu-button.has-secondary-theme {
 		&.is-style-default {
 			background: blocks-get( 'color-secondary' );
 
@@ -150,7 +150,7 @@ $scope: blocks-get( 'scope' );
 		}
 	}
 
-	&.wp-block-button.has-tertiary-theme {
+	&.wp-block-bu-button.has-tertiary-theme {
 		&.is-style-default {
 			background: blocks-get( 'color-tertiary' );
 
@@ -177,7 +177,7 @@ $scope: blocks-get( 'scope' );
 		}
 	}
 
-	&.wp-block-button.has-quaternary-theme {
+	&.wp-block-bu-button.has-quaternary-theme {
 		&.is-style-default {
 			background: blocks-get( 'color-quaternary' );
 
@@ -204,7 +204,7 @@ $scope: blocks-get( 'scope' );
 		}
 	}
 
-	&.wp-block-button.has-quinary-theme {
+	&.wp-block-bu-button.has-quinary-theme {
 		&.is-style-default {
 			background: blocks-get( 'color-quinary' );
