+++ src/blocks/button/block.json
@@ -1,0 +1,52 @@
+{
+	"$schema": "https://schemas.wp.org/trunk/block.json",
+	"apiVersion": 2,
+	"name": "bu/button",
+	"version": "0.0.3",
+	"title": "Button",
+    "description": "Prompt visitors to take action with a custom button.",
+	"category": "bu",
+	"icon": "block-default",
+	"textdomain": "bu-blocks",
+	"editorScript": "file:./index.js",
+	"editorStyle": "file:./index.css",
+	"viewStyle": "file:./style.css",
+	"render": "file:./render.php",
+    "attributes": {
+		"url": {
+			"type": "string"
+		},
+		"text": {
+			"type": "string"
+		},
+		"themeColor": {
+			"type": "string"
+		},
+		"icon": {
+			"type": "string"
+		}
+	},
+    "styles": [
+		{
+			"name": "default",
+			"label": "Default",
+			"isDefault": true
+		},
+		{
+			"name": "outline",
+			"label": "Outline"
+		},
+		{
+			"name": "text",
+			"label": "Text"
+		},
+		{
+			"name": "accent",
+			"label": "Accent"
+		}
+	],
+    "supports": {
+        "align": true,
+        "anchor": true
+	}
+}
