+++ src/blocks/button/editor.scss
@@ -5,7 +5,7 @@
  * which makes it higher in priority.
  */
 
- .wp-block-bu-button-container {
+.wp-block-bu-button-container {
 	position: relative;
 }
 
@@ -15,7 +15,7 @@
 		margin-top: 0;
 		margin-bottom: 0;
 	}
-	&.wp-block-button.wp-block-bu-button {
+	&.wp-block-bu-button {
 		color: inherit;
 
 		[data-rich-text-placeholder]:after {
